Some implemented, otis quintic has some issues
